adding bounding box

+++ Code/glmath.cpp
@@ -69,7 +69,7 @@ float dot(const vec3 &u, const vec3 &v)
 
 float Length(const vec3 &u)
 {
-	return sqrt(u.x * u.x + u.y * u.y + u.z * u.z);
+	return (float)sqrt(u.x * u.x + u.y * u.y + u.z * u.z);
 }
 
 float length2(const vec3 &u)
